+++ data/normalized/CHANGELOG.md
@@ -1,6 +1,6 @@
 # Latest Microsoft Security News Changes
 
-Generated: 2026-04-09T07:18:01.699Z
+Generated: 2026-04-10T07:23:57.905Z
 
 - 2026-07-01 • Microsoft Sentinel • Action required • Call to action: update queries and automation by July 1, 2026 - standardized account entity naming in incidents and alerts
 - 2026-07-01 • Microsoft Sentinel • GA • , Microsoft Sentinel will be supported in the Defender portal only, and any remaining customers using the Azure portal will be automatically redirected.
@@ -11,7 +11,9 @@ Generated: 2026-04-09T07:18:01.699Z
 - 2026-05-01 • Microsoft Entra • Deprecation • Plan for change – Agent Registry consolidation into Microsoft Agent 365
 - 2026-04-01 • Microsoft Defender for Cloud • GA • General availability of container security capabilities in Azure Government cloud
 - 2026-04-01 • Microsoft Defender for Endpoint • Preview • Release - macOS
+- 2026-04-01 • Microsoft Defender for Endpoint • Ikke oppgitt • Release version 20.126022.20.0 released: See
 - 2026-04-01 • Microsoft Defender for Endpoint • Ikke oppgitt • Release version 20.126022.18.0 released: See
+- 2026-04-01 • Microsoft Defender for Endpoint • Ikke oppgitt • Build 1.0.8805.0103 released: See
 - 2026-04-01 • Microsoft Defender XDR • Ikke oppgitt • Microsoft Defender Experts for XDR customers can now see
 - 2026-04-01 • Microsoft Purview • GA • General availability (GA)
 - 2026-04-01 • Microsoft Sentinel • GA • Microsoft Sentinel data federation (Preview)
@@ -20,5 +22,3 @@ Generated: 2026-04-09T07:18:01.699Z
 - 2026-03-22 • Microsoft Defender for Cloud • Preview • Expanded multicloud coverage for AWS and GCP (Preview)
 - 2026-03-11 • Microsoft Defender for Cloud • GA • Kubernetes gated deployment support for AKS Automatic (GA)
 - 2026-03-10 • Microsoft Defender for Cloud • Preview • Code to runtime enrichment for recommendations
-- 2026-03-10 • Microsoft Defender for Cloud • Ikke oppgitt • Severity‑based risk assignment for "Not evaluated" recommendations
-- 2026-03-04 • Microsoft Defender for Cloud • Deprecation • Deprecation of preview of container and container images vulnerability recommendations